+++ pages/learn/courses/_code/assessments/index.vue
@@ -1,15 +1,15 @@
 <template>
   <ApolloQuery
-    v-slot="{ isLoading, result: { data: course, error } }"
+    v-slot="{ isLoading, result: { error } }"
     :query="require('~/gql/learn/getCourse.gql')"
     :update="(data) => data.course"
     :variables="{ code: courseCode }"
-    @result="setTitle"
+    @result="setResult"
   >
     <page-title :loading="!!isLoading" :value="title" />
 
-    <div v-if="!error">
-      <v-row v-if="course?.isRegistered">
+    <div v-if="!error && canShowContent">
+      <v-row>
         <v-col cols="12" md="9">
           <v-card>
             <v-tabs v-model="currentTab" show-arrows>
@@ -62,10 +62,11 @@
 import titles from '@/mixins/titles.js'
 
 export default {
-  name: 'LearnCourseAssessmentsPage',
+  name: 'LearnAssessmentsPage',
   mixins: [titles],
   data() {
     return {
+      course: null,
       currentTab: 0,
       title: '',
     }
@@ -76,12 +77,16 @@ export default {
     }
   },
   computed: {
+    canShowContent() {
+      return !this.course || this.course.isRegistered
+    },
     courseCode() {
       return this.$route.params.code
     },
   },
   methods: {
-    setTitle({ data: course }) {
+    setResult({ data: course }) {
+      this.course = course
       this.title = course?.name ?? ''
     },
   },